blogapi/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
djangorestframework==3.16.1
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
packaging==26.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
